+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - Calendar Events nudge [#22](https://github.com/rokwire/lms-building-block/issues/22)
 
+### Changed
+- Early completion nudge update [#28](https://github.com/rokwire/lms-building-block/issues/28)
+
 ### Added
 - Prepare the project to become open source [#5](https://github.com/rokwire/lms-building-block/issues/5)
 - Deep links in nudges [#26](https://github.com/rokwire/lms-building-block/issues/26)